task 6

+++ 0x00-hello_world/6-size.c
@@ -2,7 +2,6 @@
 #include <math.h>
 /**
  * main - Entry point
- *
  * Return: Always 0 (Success)
  */
 int main(void)
@@ -12,11 +11,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-
-	printf("Size of a char: %d byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %d byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(c));
-	rintf("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %d byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a char: %d byte(s)\n", sizeof(a));
+	printf("Size of an int: %d byte(s)\n", sizeof(b));
+	printf("Size of a long int: %d byte(s)\n", sizeof(c));
+	rintf("Size of a long long int: %d byte(s)\n", sizeof(d));
+	printf("Size of a float: %d byte(s)\n", sizeof(e));
 	return (0);
 }
